model_split
